add libretro lib to support path_mkdir
