+++ src/osd/retro/libretro-common/include/retro_common_api.h
@@ -1,4 +1,4 @@
-/* Copyright  (C) 2010-2016 The RetroArch team
+/* Copyright  (C) 2010-2017 The RetroArch team
  *
  * ---------------------------------------------------------------------------------------
  * The following license statement only applies to this file (retro_common_api.h).
@@ -79,7 +79,7 @@ typedef int ssize_t;
 #define STRING_REP_INT64  "%I64u"
 #define STRING_REP_UINT64 "%I64u"
 #define STRING_REP_ULONG  "%Iu"
-#elif defined(__STDC_VERSION__) && __STDC_VERSION__>=199901L
+#elif defined(__STDC_VERSION__) && __STDC_VERSION__>=199901L && !defined(VITA) && !defined(WIIU)
 #define STRING_REP_INT64  "%llu"
 #define STRING_REP_UINT64 "%llu"
 #define STRING_REP_ULONG  "%zu"
